# contracts/openapi/seekflux-v1.yaml
openapi: 3.1.0
info:
  title: SeekFlux API
  version: 0.1.0
  description: SeekFlux content, search, recommendation, Agent Runtime, and interaction contracts.
servers:
  - url: http://localhost:8080
    description: Direct Search and recommendation
  - url: http://localhost:8083
    description: Agent Server
paths:
  /v1/agent/search:
    post:
      operationId: agentSearch
      summary: Execute the bounded Search Agent and fall back to Direct Search when required
      tags: [Agent]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentSearchRequest"
      responses:
        "200":
          description: Stable Agent terminal state, results, and correlated Agent/Search traces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/AgentConflict"
        "503":
          $ref: "#/components/responses/AgentUnavailable"
  /v1/agent/sessions/{sessionId}:cancel:
    post:
      operationId: cancelAgentSession
      summary: Request cancellation of the locally executing Agent Session
      tags: [Agent]
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
      responses:
        "200":
          description: Whether an in-process execution received the cancellation signal
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, cancelled]
                properties:
                  sessionId: {type: string}
                  cancelled: {type: boolean}
...
  /v1/search:
    get:
      operationId: searchByQueryParameter
      summary: Search published content profiles
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          schema: {type: string, minLength: 1, maxLength: 500}
        - name: page
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          schema: {type: integer, minimum: 1, maximum: 50, default: 12}
        - name: required_tags
          in: query
          description: Optional repeated exact tags that every returned content must contain.
          schema:
            type: array
            maxItems: 10
            items: {type: string, maxLength: 64}
      responses:
        "200":
          description: Hybrid Direct Search result page with retrieval trace and degradation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResultPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/SearchUnavailable"
    post:
      operationId: search
      summary: Search published content profiles
      tags: [Search]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search result page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResultPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/SearchUnavailable"
...
components:
...
  schemas:
    AgentSearchOptions:
      type: object
      properties:
        allowClarification: {type: boolean, default: true}
    AgentSearchRequest:
      type: object
      required: [sessionId, turnId, query]
      properties:
        requestId: {type: string, maxLength: 128}
        sessionId: {type: string, minLength: 1, maxLength: 128}
        turnId: {type: string, minLength: 1, maxLength: 128}
        agentId:
          type: string
          maxLength: 128
          default: search-assistant
          enum: [search-assistant, search-precise]
        mode:
          type: string
          default: AUTO
          enum: [AUTO, DIRECT, AGENT]
          description: AUTO sends simple queries directly to Search and complex or multi-turn queries to the Agent.
        query: {type: string, minLength: 1, maxLength: 500}
        page: {type: integer, minimum: 0, default: 0}
        size: {type: integer, minimum: 1, maximum: 50, default: 12}
        requiredTags:
          type: array
          maxItems: 10
          items: {type: string, maxLength: 64}
        constraintPatch:
          $ref: "#/components/schemas/ConstraintPatchRequest"
        options:
          $ref: "#/components/schemas/AgentSearchOptions"
    ConstraintPatchRequest:
      type: object
      required: [baseVersion]
      description: Optimistic, version-checked update to the persisted Search Goal.
      properties:
        baseVersion: {type: integer, format: int64, minimum: 1}
        replacementQuery: {type: string, minLength: 1, maxLength: 500}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 50}
        addRequiredTags:
          type: array
          maxItems: 10
          items: {type: string, maxLength: 64}
        removeRequiredTags:
          type: array
          maxItems: 10
          items: {type: string, maxLength: 64}
    QueryConstraintSet:
      type: object
      required: [page, size, requiredTags]
      properties:
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 50}
        requiredTags:
          type: array
          maxItems: 10
          items: {type: string, maxLength: 64}
    SearchPlan:
      type: object
      required: [originalQuery, rewrittenQuery, derivedRequiredTags, complex, reasons]
      properties:
        originalQuery: {type: string}
        rewrittenQuery: {type: string}
        derivedRequiredTags:
          type: array
          items: {type: string}
        complex: {type: boolean}
        reasons:
          type: array
          items:
            type: string
            enum: [LONG_NATURAL_LANGUAGE_QUERY, EXPLICIT_CONSTRAINT_LANGUAGE, MULTI_SLOT_QUERY, MULTI_TURN_GOAL]
    AgentStepTrace:
      type: object
      required: [step, action, status, tookMillis]
      properties:
        step: {type: integer, minimum: 1}
        action: {type: string, enum: [PLAN, CALL_TOOL, COMPLETE, CLARIFY, FALLBACK]}
        status: {type: string}
        toolCallId: {type: [string, "null"]}
        toolName: {type: [string, "null"]}
        linkedTraceId: {type: [string, "null"]}
        tookMillis: {type: integer, format: int64, minimum: 0}
        errorCode: {type: [string, "null"]}
    AgentTrace:
      type: object
      required:
        - agentRunId
        - agentId
        - agentVersion
        - plannerVersion
        - promptVersion
        - decisionProviderVersion
        - toolSchemaVersions
        - startedAt
        - tookMillis
        - terminalState
        - executionMode
        - steps
      properties:
        agentRunId: {type: string, format: uuid}
        agentId: {type: string}
        agentVersion: {type: string}
        plannerVersion: {type: string}
        promptVersion: {type: string}
        decisionProviderVersion: {type: string}
        toolSchemaVersions:
          type: object
          additionalProperties: {type: string}
        startedAt: {type: string, format: date-time}
        tookMillis: {type: integer, format: int64, minimum: 0}
        terminalState:
          type: string
          enum: [RESULTS_READY, NEED_CLARIFICATION, FALLBACK_REQUIRED, CANCELLED, FAILED]
        executionMode:
          type: string
          enum: [AGENT, AGENT_TO_DIRECT_FALLBACK]
        fallbackReason: {type: [string, "null"]}
        steps:
          type: array
          items:
            $ref: "#/components/schemas/AgentStepTrace"
    AgentSearchResponse:
      type: object
      required:
        - requestId
        - agentRunId
        - sessionId
        - turnId
        - state
        - executionMode
        - goalVersion
        - routeReason
        - searchPlan
        - appliedConstraints
        - total
        - page
        - size
        - items
        - agentTrace
        - successfulToolCount
        - candidateSetReused
        - degraded
      properties:
        requestId: {type: string}
        agentRunId: {type: [string, "null"], format: uuid}
        sessionId: {type: string}
        turnId: {type: string}
        state:
          type: string
          enum: [RESULTS_READY, NEED_CLARIFICATION, FALLBACK_RESULTS, CANCELLED, FAILED]
        executionMode:
          type: string
          enum: [DIRECT, AGENT, AGENT_TO_DIRECT_FALLBACK]
        goalVersion:
          type: integer
          format: int64
          minimum: 0
          description: Zero for stateless Direct routing; positive for a committed Agent Search Goal.
        routeReason:
          type: string
          enum: [SIMPLE_QUERY, COMPLEX_QUERY, MULTI_TURN_PATCH, EXPLICIT_DIRECT, EXPLICIT_AGENT]
        searchPlan:
          $ref: "#/components/schemas/SearchPlan"
        appliedConstraints:
          $ref: "#/components/schemas/QueryConstraintSet"
        clarification: {type: [string, "null"]}
        total: {type: integer, format: int64, minimum: 0}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 50}
        items:
          type: array
          items:
            $ref: "#/components/schemas/SearchHit"
        searchTrace:
          oneOf:
            - $ref: "#/components/schemas/SearchTrace"
            - type: "null"
        agentTrace:
          oneOf:
            - $ref: "#/components/schemas/AgentTrace"
            - type: "null"
        selectedTool: {type: [string, "null"]}
        successfulToolCount: {type: integer, minimum: 0}
        candidateSetReused:
          type: boolean
          description: True when the response returns one Search Tool candidate set without Agent-side reranking.
        degraded: {type: boolean}
        fallbackReason: {type: [string, "null"]}
...
    SearchRequest:
      type: object
      required: [query]
      properties:
        query: {type: string, minLength: 1, maxLength: 500}
        page: {type: integer, minimum: 0, default: 0}
        size: {type: integer, minimum: 1, maximum: 50, default: 12}
        requiredTags:
          type: array
          maxItems: 10
          items: {type: string, maxLength: 64}
    SearchHit:
      type: object
      required: [contentId, creatorId, mediaUri, title, description, summary, tags, profileVersion, score, sources, publishedAt]
      properties:
        contentId: {type: string, format: uuid}
        creatorId: {type: string}
        mediaUri: {type: string}
        title: {type: string}
        description: {type: string}
        summary: {type: string}
        tags:
          type: array
          items: {type: string}
        profileVersion: {type: integer, minimum: 1}
        score: {type: number, format: double}
        sources:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [KEYWORD, SEMANTIC]
        publishedAt: {type: string, format: date-time}
    SearchResultPage:
      type: object
      required: [query, total, page, size, tookMillis, hits, trace]
      properties:
        query: {type: string}
        total: {type: integer, format: int64, minimum: 0}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1}
        tookMillis: {type: integer, format: int64, minimum: 0}
        hits:
          type: array
          items:
            $ref: "#/components/schemas/SearchHit"
        trace:
          $ref: "#/components/schemas/SearchTrace"
    SearchChannelTrace:
      type: object
      required: [source, status, retrieverVersion, tookMillis, candidateCount]
      properties:
        source:
          type: string
          enum: [KEYWORD, SEMANTIC]
        status:
          type: string
          enum: [SUCCESS, FAILED, TIMED_OUT, REJECTED]
        retrieverVersion: {type: string}
        tookMillis: {type: integer, format: int64, minimum: 0}
        candidateCount: {type: integer, minimum: 0}
        errorCode:
          oneOf:
            - type: string
            - type: "null"
    SearchTrace:
      type: object
      required:
        - requestId
        - executionMode
        - indexVersion
        - policyVersion
        - tookMillis
        - degraded
        - unavailableSources
        - channels
      properties:
        requestId: {type: string}
        executionMode:
          type: string
          enum: [DIRECT_HYBRID, DIRECT_KEYWORD_FALLBACK, DIRECT_SEMANTIC_FALLBACK]
        indexVersion: {type: string}
        policyVersion: {type: string}
        tookMillis: {type: integer, format: int64, minimum: 0}
        degraded: {type: boolean}
        unavailableSources:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [KEYWORD, SEMANTIC]
        channels:
          type: array
          items:
            $ref: "#/components/schemas/SearchChannelTrace"
...
    ApiError:
      type: object
      required: [code, message, timestamp]
      properties:
        code: {type: string}
        message: {type: string}
        timestamp: {type: string, format: date-time}
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    SearchUnavailable:
      description: All Direct Search retrieval channels failed or exceeded the request deadline
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    AgentConflict:
      description: Agent Session is busy, the request id was already committed, or a Search Goal patch used a stale version
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    AgentUnavailable:
      description: Agent Runtime, Session store, or execution authority is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"